rename gwells.js to ApiService.js
